\name{bSims-package}
\alias{bSims-package}
\alias{bSims}
\docType{package}
\title{
\packageTitle{bSims}
}
\description{
\packageDescription{bSims}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bSims}
\packageIndices{bSims}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bSims}

Maintainer: \packageMaintainer{bSims}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
