\name{bSims-package}
\alias{bSims-package}
\alias{bSims}
\docType{package}
\title{
\packageTitle{bSims}
}
\description{
\packageDescription{bSims}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bSims}
\packageIndices{bSims}
}
\author{
\packageAuthor{bSims}

Maintainer: \packageMaintainer{bSims}
}
\keyword{ package }
